requirements.txt
anthropic>=0.42.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.0
